mensageiroBackend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pika==1.3.2
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
daphne==4.0.0 